Add 2DMaze
Added 2DMaze. Currently need to rework validation, as it
currently needs a maze solver to confirm that it's solvable.

Also, need to implement optparse-applicative for this project,
as the code for taking and displaying options is taking
up a lot of space.

The maze parser could also be rendered more concise, since
resorting to foldr is somewhat of an ugly solution.

There is also potential space in adding support for online
loading of the web repository of mazes.

In theory, this can also be ported to Brick, but currently
the game should, at least in theory, work on Windows,
which Brick isn't supporting as of yet.