Added Clickbait Headline Generator, one partial
I added a complete translation of #11, clickbait
headline generator, as well as a partial version of
number 77, Tower of Hanoi.

The CHG is interesting because it is about 20% shorter
than the Python original, but that's partially due
to using better idioms.

The Tower of Hanoi is incomplete because it still cannot
display the Hanoi datatype.