Rewrite Clickbait to be more idiomatic
Rewrote Clickbait to be more idiomatic by using a
custom state type with support for stdgen. Issues with
superfluous pures were handled by turning strings into
lambdas written in a list of strings style, using fold
as needed to compress the list.

Still not written for readability, but the code is of
reasonable quality.