Finish Tower of Hanoi example.
I finished the Tower of Hanoi example, so this
commit introduces the Tower of Hanoi program as complete.

In contrast to the Python original, this includes color via
ansi-terminal, which is likely slowing the program down, and
it also features a key-press based interface, which may
remove its ability to be cross platform, since the
keys are based on POSIX, not Windows.